Split hooks/utils to separate files
